last step: add scheduling

+++ app/Http/Controllers/API/LessonController.php
@@ -12,9 +12,13 @@ class LessonController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function __construct() {
+        $this->middleware('auth:api')->except(['index', 'show']);
+    }
+    public function index(Request $request)
     {
-        $lesson = LessonResource::collection(Lesson::all());
+        $limit = $request->input('limit') <= 50 ? $request->input('limit') : 15;
+        $lesson = LessonResource::collection(Lesson::paginate($limit));
         return $lesson->response()->setStatusCode(200);
     }
 
@@ -41,6 +45,8 @@ class LessonController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $idLesson = Lesson::findOrFail($id);
+        $this->authorize('delete', $idLesson);
         $lesson = new LessonResource(Lesson::findOrFail($id));
         $lesson->update($request->all());                                   
     
@@ -53,6 +59,8 @@ class LessonController extends Controller
      */
     public function destroy($id)
     {
+        $lesson = Lesson::findOrFail($id);
+        $this->authorize('delete', $lesson);
         Lesson::findOrFail($id)->delete();
     
         return 204;
