+++ app/Http/Controllers/API/UserController.php
@@ -13,9 +13,13 @@ class UserController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function __construct() {
+        $this->middleware('auth:api')->except(['index', 'show']);
+    }
+    public function index(Request $request)
     {
-        $user = UserResource::collection(User::all());
+        $limit = $request->input('limit') <= 50 ? $request->input('limit') : 15;
+        $user = UserResource::collection(User::paginate($limit));
         return $user->response()->setStatusCode(200);
     }
 
@@ -24,6 +28,7 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', User::class);
         $user = new UserResource(User::create([
             'name' => $request->name,
             'email' => $request->email,
@@ -46,6 +51,8 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $idUser = User::findOrFail($id);
+        $this->authorize('update', $idUser);
         $user = new UserResource(User::findOrFail($id));
 
         $user->update($request->all());
@@ -58,6 +65,8 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $idUser = User::findOrFail($id);
+        $this->authorize('delete', $idUser);
         User::findOrFail($id)->delete();
         
         return 204;
